fix: reorganizando as models e corrigindo os erros
