+++ src/app/models/Aluno.ts
@@ -1,0 +1,15 @@
+import { Usuario } from "./Usuario";
+import { Projeto } from "./Projeto";
+
+export interface Aluno {
+    id: number;
+    usuario?: Usuario;
+    projeto?: Projeto;
+    foto_url: string;
+    data_criacao: Date;
+    data_alteracao: Date;
+    data_ingresso: Date;
+    curso_semestre: string;
+  }
+
+
