+++ src/app/models/Evento.ts
@@ -1,0 +1,15 @@
+import { Candidato } from "./Candidato";
+
+export interface Evento {
+    id: number;
+    nome: string;
+    tipo_evento: string;
+    data_inicio: Date;
+    data_fim: Date;
+    descricao: string;
+    status: string;
+    data_alteracao: Date;
+    data_criacao: Date;
+    curso_semestre: string;
+    candidatos: Candidato[];
+  }
