--- src/app/models/candidatos.ts
@@ -1,9 +1,0 @@
-export interface Candidato {
-    id: number;
-    id_aluno?: number;
-    id_projeto?: number;
-    id_evento?: number;
-    situacao_candidato: string;
-    qrcode: string;
-  }
-
