--- src/app/models/eventos.ts
@@ -1,12 +1,0 @@
-export interface Evento {
-    id: number;
-    nome: string;
-    tipo_evento: string;
-    data_inicio: Date;
-    data_fim: Date;
-    descricao: string;
-    status: string;
-    data_alteracao: Date;
-    data_criacao: Date;
-    curso_semestre: string;
-  }
